requirements.txt
requests
beautifulsoup4
ipython
django
djangorestframework
django-cors-headers
jedi
funcy
retry
